refactor: put all database methods + tests in their own folder
