+++ database/delete-rotation/delete-rotation.test.ts
@@ -1,7 +1,7 @@
 import { GetCommand } from "@aws-sdk/lib-dynamodb";
 import { deleteRotation, MISSING_DELETE_PARAMS } from "./delete-rotation";
-import { getDynamoDBClient } from "./get-dynamodb-client";
-import seed from "../seed/rotations.json";
+import { getDynamoDBClient } from "../get-dynamodb-client";
+import seed from "../../seed/rotations.json";
 
 const ddb = getDynamoDBClient();
