+++ database/get-all-rotations/get-all-rotations.test.ts
@@ -1,4 +1,4 @@
-import rotations from "../seed/rotations.json";
+import rotations from "../../seed/rotations.json";
 import { getAllRotations } from "./get-all-rotations";
 
 it("should get all rotations in table", async () => {
