+++ database/get-all-rotations/get-all-rotations.ts
@@ -1,6 +1,6 @@
-import { getDynamoDBClient } from "./get-dynamodb-client";
+import { getDynamoDBClient } from "../get-dynamodb-client";
 import { ScanCommand } from "@aws-sdk/lib-dynamodb";
-import { Rotation } from "../entities/rotation";
+import { Rotation } from "../../entities/rotation";
 
 const dynamodb = getDynamoDBClient();
