+++ database/get-all-rotations/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-all-rotations";
