+++ database/get-rotations-by-channel-id/get-rotations-by-channel-id.ts
@@ -1,6 +1,6 @@
-import { getDynamoDBClient } from "./get-dynamodb-client";
+import { getDynamoDBClient } from "../get-dynamodb-client";
 import { QueryCommand } from "@aws-sdk/lib-dynamodb";
-import { Rotation } from "../entities/rotation";
+import { Rotation } from "../../entities/rotation";
 
 const dynamodb = getDynamoDBClient();
