+++ database/get-rotations-by-channel-id/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-rotations-by-channel-id";
