+++ database/put-rotation/put-rotation.ts
@@ -1,9 +1,9 @@
 import dayjs from "dayjs";
 import { v4 as uuid } from "uuid";
-import { PutCommand } from "@aws-sdk/lib-dynamodb"; // ES6 import
+import { PutCommand } from "@aws-sdk/lib-dynamodb";
 
-import { Rotation } from "../entities/rotation";
-import { getDynamoDBClient } from "./get-dynamodb-client";
+import { Rotation } from "../../entities/rotation";
+import { getDynamoDBClient } from "../get-dynamodb-client";
 
 const dynamodb = getDynamoDBClient();
